backend/requirements.txt
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser
gunicorn==20.1.0
python-dotenv==1.0.1
PyYAML==6.0
Pillow==11.0.0
drf-extra-fields==3.7.0
psycopg2-binary==2.9.10
webcolors==1.11.1
reportlab==4.2.2
